requirements.txt
pylast
winsdk
customtkinter
Pillow
asyncio

python-dotenv
